requirements.txt

streamlit>=1.36.0
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.6.0
